Validate exceptions on tokens thru the Endpoint
The endpoint integration tests are racing the late collection
behaviour in the profiler endpoint, which sometimes results in
tests failing due to receiving outdated profile information.

We are attempting to correct this by fetching the profile data
by its token thru another endpoint request. Hopefully this will
slow things down enough to give the late write a much better chance
to finish before the data is requested.

In practice this isn't an actual problem, as in most cases (like the
Dashboard) we poll for updated information.